refactor: relocate module repositories into repositories/ subfolder
