+++ src/server/api/modules/clubs/usecases/admin-get-club-images-upload-url.usecase.ts
@@ -3,7 +3,7 @@ import {
   AdminGetClubImagesUploadUrlOutputDTOSchema,
   type AdminGetClubImagesUploadUrlInputDTO,
 } from "@/server/api/modules/clubs/dto";
-import { clubsRepository } from "@/server/api/modules/clubs/clubs.repository";
+import { clubsRepository } from "@/server/api/modules/clubs/repositories/clubs.repository";
 import { getExtension, getPublicUrl, getSignedUploadUrl } from "@/server/services/r2";
 import { notFound } from "@/server/errors";
