+++ src/server/api/modules/clubs/usecases/admin-update-club.usecase.ts
@@ -1,6 +1,6 @@
-import { clubsRepository } from "@/server/api/modules/clubs/clubs.repository";
-import { usersRepository } from "@/server/api/modules/users/users.repository";
-import { clubCategoriesRepository } from "@/server/api/modules/clubs/club-categories.repository";
+import { clubsRepository } from "@/server/api/modules/clubs/repositories/clubs.repository";
+import { usersRepository } from "@/server/api/modules/users/repositories/users.repository";
+import { clubCategoriesRepository } from "@/server/api/modules/clubs/repositories/club-categories.repository";
 import { unitOfWork } from "@/server/db/unit-of-work";
 import { notFound } from "@/server/errors";
 import type {
