+++ src/server/api/modules/clubs/usecases/get-all-clubs-admin.usecase.ts
@@ -1,4 +1,4 @@
-import { clubsRepository } from "@/server/api/modules/clubs/clubs.repository";
+import { clubsRepository } from "@/server/api/modules/clubs/repositories/clubs.repository";
 import type {
   GetAllClubsAdminInputDTO,
   GetAllClubsAdminOutputDTO,
