+++ src/server/api/modules/clubs/usecases/get-club-logo-upload-url.usecase.ts
@@ -2,7 +2,7 @@ import {
   GetClubLogoUploadUrlOutputDTOSchema,
   type GetClubLogoUploadUrlInputDTO,
 } from "@/server/api/modules/clubs/dto";
-import { clubsRepository } from "@/server/api/modules/clubs/clubs.repository";
+import { clubsRepository } from "@/server/api/modules/clubs/repositories/clubs.repository";
 import { getExtension, getPublicUrl, getSignedUploadUrl } from "@/server/services/r2";
 import { notFound, validationError } from "@/server/errors";
