+++ src/server/api/modules/clubs/usecases/submit-club-profile-registration.usecase.ts
@@ -1,11 +1,11 @@
-import { clubsRepository } from "@/server/api/modules/clubs/clubs.repository";
+import { clubsRepository } from "@/server/api/modules/clubs/repositories/clubs.repository";
 import {
   SubmitClubProfileRegistrationOutputDTOSchema,
   type SubmitClubProfileRegistrationInputDTO,
   type SubmitClubProfileRegistrationOutputDTO,
 } from "@/server/api/modules/clubs/dto";
 import { validationError, notFound } from "@/server/errors";
-import { usersRepository } from "@/server/api/modules/users/users.repository";
+import { usersRepository } from "@/server/api/modules/users/repositories/users.repository";
 import { unitOfWork } from "@/server/db/unit-of-work";
 
 export const submitClubProfileRegistration = async (
