+++ src/server/api/modules/master-data/usecases/get-all-activity-types.usecase.ts
@@ -1,4 +1,4 @@
-import { masterDataRepository } from "@/server/api/modules/master-data/master-data.repository";
+import { masterDataRepository } from "@/server/api/modules/master-data/repositories/master-data.repository";
 import type { GetAllActivityTypesOutputDTO } from "@/server/api/modules/master-data/dto";
 
 export const getAllActivityTypes = async (): Promise<GetAllActivityTypesOutputDTO> => {
