+++ src/server/api/modules/master-data/usecases/update-affiliation.usecase.ts
@@ -1,4 +1,4 @@
-import { masterDataRepository } from "@/server/api/modules/master-data/master-data.repository";
+import { masterDataRepository } from "@/server/api/modules/master-data/repositories/master-data.repository";
 import type {
   UpdateAffiliationInputDTO,
   UpdateAffiliationOutputDTO,
